Applying ACL  to both controller

+++ app/Http/Controllers/CategoriesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Category;
 use App\Http\Requests\CreateCategoryRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoriesController extends Controller
 {
@@ -28,7 +29,15 @@ class CategoriesController extends Controller
 
     public  function  store(CreateCategoryRequest $request)
     {
+
       //dd($request);
+
+        // Appy ACL
+        if( Auth::user()->role == 3)
+        {
+            return redirect()->back()->with('error', 'Access denied you dont have enough sufficient privileges');
+        }
+
         $category = new Category($request->all());
         $category->save();
 
@@ -36,6 +45,12 @@ class CategoriesController extends Controller
     }
     public function  edit($id)
     {
+        // Appy ACL
+        if( Auth::user()->role == 3)
+        {
+            return redirect()->back()->with('error', 'Access denied you dont have enough sufficient privileges');
+        }
+
        $data['category'] = $this->categories->where('id', $id)->first(); //dd($data['category']);
 
        return view('categories.edit-category',$data);
@@ -45,6 +60,13 @@ class CategoriesController extends Controller
 
     public function  update(CreateCategoryRequest $request , $id)
     {
+        // Appy ACL
+        if(Auth::user()->role == 3)
+        {
+            return redirect()->back()->with('error', 'Access denied you dont have enough sufficient privileges');
+        }
+
+
        $category = $this->categories->where('id', $id)->first();
        $category->name = $request ->name;
        $category->save();
@@ -55,6 +77,13 @@ class CategoriesController extends Controller
 
     public function  delete($id)
     {
+        // Appy ACL
+        if(Auth::user()->role == 2 || Auth::user()->role == 3)
+        {
+            return redirect()->back()->with('error', 'Access denied you dont have enough sufficient privileges');
+        }
+
+
         $category = $this->categories->where('id', $id);
 
         $category->delete();
